feat: freezer service
