+++ util/types/schemas/blockchain.mol
@@ -98,6 +98,12 @@ table Block {
     proposals:              ProposalShortIdVec,
 }
 
+table ArchivedBlock {
+    uncles:                 UncleBlockVec,
+    transactions:           TransactionVec,
+    proposals:              ProposalShortIdVec,
+}
+
 table CellbaseWitness {
     lock:    Script,
     message: Bytes,
